move document handlers into folder so we can split it later
